WinGit: verify that documentation is checked out with autocrlf=true
The documentation contain txt files that should contain
the correct line endings on Windows. Therefore they must
be checked out using newline conversion.

This is now checked before copying files.

+++ share/WinGit/copy-files.sh
@@ -7,6 +7,17 @@ test -z "$1" && {
 
 test -d /doc/git/html/.git || { echo "Error: html pages in /doc/git/html/.git missing"; exit 1; }
 
+if [ "$( cd /doc/git/html/ ; git config core.autocrlf )" != "true" ]
+then 
+	echo "Error: documentation must be checked out with core.autocrlf=true."
+	echo "Hint: fix this by"
+	echo "   cd /doc/git/html"
+	echo "   git config core.autocrlf true"
+	echo "   rm *.html *.txt howto"
+	echo "   git checkout -f"
+	exit 1
+fi
+
 TMPDIR=$1
 
 (test ! -d "$TMPDIR" || echo "Removing $TMPDIR" && rm -rf "$TMPDIR") &&
